+++ src/main/kotlin/dev/aaronhowser/mods/aaron/misc/AaronExtensions.kt
@@ -139,6 +139,9 @@ object AaronExtensions {
 	fun RandomSource.nextRange(min: Int, max: Int): Int = nextInt(max - min) + min
 	fun RandomSource.chance(chance: Number): Boolean = nextDouble() <= chance.toDouble()
 
+	fun RandomSource.roll(chance: Number): Boolean = nextDouble() <= chance.toDouble()
+	fun RandomSource.rollDie(sides: Int): Boolean = nextInt(0, sides) == 0
+
 	fun Number.toDegrees(): Double = Math.toDegrees(this.toDouble())
 	fun Number.toRadians(): Double = Math.toRadians(this.toDouble())